health_analyzer_requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Frontend Interface
streamlit==1.31.0
gradio==4.16.0

# Database
mysql-connector-python==8.3.0
pymysql==1.1.0
SQLAlchemy==2.0.25

# Machine Learning
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# AI/NLP Models
transformers==4.37.2
torch==2.2.0
sentence-transformers==2.3.1
huggingface-hub==0.20.3

# Document Processing
PyMuPDF==1.23.21
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
aiofiles==23.2.1

# Date/Time handling
python-dateutil==2.8.2

# Logging and monitoring
loguru==0.7.2

# Testing (optional)
pytest==8.0.0
pytest-asyncio==0.23.3

# Security
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2